# docs/swagger.yaml
definitions:
...
  db.Sign:
    properties:
      sign_id:
        type: integer
      sign_ip:
        type: string
      sign_name:
        type: object
      sign_password:
        type: string
      sign_port:
        type: integer
      sign_type:
        type: string
      sign_username:
        type: string
      zone_id:
        type: integer
    required:
    - sign_id
    - sign_ip
    - sign_name
    - sign_password
    - sign_port
    - sign_type
    - sign_username
    - zone_id
    type: object
...
info:
  contact: {}
  title: Find Your Car
  version: 0.0.9
paths:
...
  /fyc/sign:
    delete:
      description: Soft delete a sign by setting the is_deleted flag to true
      parameters:
      - description: sign ID
        in: query
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: sign deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid sign ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete sign
          schema:
            additionalProperties: true
            type: object
      summary: Soft delete a sign
      tags:
      - Sign
    get:
      description: Get a list of signs (all, enabled, or deleted) or a specific sign
        by ID with optional extra data
      parameters:
      - description: sign ID
        in: query
        name: id
        type: integer
      - description: 'Type of sign: all, enabled, or deleted (default is ''all'')'
        in: query
        name: type
        type: string
      - description: Include extra information if 'yes'
        in: query
        name: extra
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of signs or a single sign
          schema:
            $ref: '#/definitions/db.Sign'
        "400":
          description: 'Bad request: Invalid sign ID'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: No sign found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get signs or a specific sign by ID
      tags:
      - Sign
    post:
      consumes:
      - application/json
      description: Add a new sign to the database
      parameters:
      - description: sign data
        in: body
        name: sign
        required: true
        schema:
          $ref: '#/definitions/db.Sign'
      produces:
      - application/json
      responses:
        "201":
          description: sign created successfully
          schema:
            $ref: '#/definitions/db.Sign'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create a new sign
          schema:
            additionalProperties: true
            type: object
      summary: Add a new sign
      tags:
      - Sign
    put:
      consumes:
      - application/json
      description: Update an existing sign by ID
      parameters:
      - description: sign ID
        in: query
        name: id
        required: true
        type: integer
      - description: Updated sign data
        in: body
        name: sign
        required: true
        schema:
          $ref: '#/definitions/db.Sign'
      produces:
      - application/json
      responses:
        "200":
          description: sign updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload or ID mismatch
          schema:
            additionalProperties: true
            type: object
        "404":
          description: sign not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update sign
          schema:
            additionalProperties: true
            type: object
      summary: Update a sign by ID
      tags:
      - Sign
...
securityDefinitions:
  BearerAuth3rdParty:
    description: 'Authorization token for third-party section (Ensure the token is
      in this format: Bearer token)'
    in: header
    name: Authorization
    type: apiKey
  BearerAuthBackOffice:
    description: 'Authorization token for back-office section (Ensure the token is
      in this format: Bearer token)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"